BacnetWritePriority enum in a human readable form, and usage (settings, ...)

+++ Yabe/MainDialog.cs
@@ -1455,7 +1455,7 @@ namespace Yabe
                 //write
                 try
                 {
-                    comm.WritePriority = Properties.Settings.Default.DefaultWritePriority;
+                    comm.WritePriority = (uint)Properties.Settings.Default.DefaultWritePriority;
                     if (!comm.WritePropertyRequest(adr, object_id, (BacnetPropertyIds)property.propertyIdentifier, b_value))
                     {
                         MessageBox.Show(this, "Couldn't write property", "Communication Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
